Xione (Broadcom) platform - Added CMA and GPU memory metrics

+++ PerformanceMetric.cpp
@@ -79,8 +79,8 @@ void PerformanceMetric::CollectData(std::chrono::seconds frequency)
         GetAvSessionStats();
 
         auto end = std::chrono::high_resolution_clock::now();
-        LOG_INFO("PerformanceMetric completed in %ld us",
-                 std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
+        LOG_INFO("PerformanceMetric completed in %lld us",
+                 (long long)std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
 
         // Wait for period before doing collection again, or until cancelled
         mCv.wait_for(lock, frequency);
